🚚️ api: move model to src/enrolment context
